@@ -1,14 +1,9 @@
-﻿using Microsoft.EntityFrameworkCore.Storage;
-using System.Diagnostics.CodeAnalysis;
-using System.Text;
+﻿namespace gRPCServer.Models;
 
-namespace gRPCServer.Models
+public class CustomNameSqlGenerationHelper : RelationalSqlGenerationHelper
 {
-    public class CustomNameSqlGenerationHelper : RelationalSqlGenerationHelper
-    {
-        private static string Customize(string input) => input.ToLower();
-        public CustomNameSqlGenerationHelper([NotNull] RelationalSqlGenerationHelperDependencies dependencies) :base(dependencies) { }
-        public override string DelimitIdentifier(string identifier) => base.DelimitIdentifier(Customize(identifier));
-        public override void DelimitIdentifier(StringBuilder builder,string identifier) => base.DelimitIdentifier(builder,Customize(identifier));
-    }
+    private static string Customize(string input) => input.ToLower();
+    public CustomNameSqlGenerationHelper([NotNull] RelationalSqlGenerationHelperDependencies dependencies) : base(dependencies) { }
+    public override string DelimitIdentifier(string identifier) => base.DelimitIdentifier(Customize(identifier));
+    public override void DelimitIdentifier(StringBuilder builder, string identifier) => base.DelimitIdentifier(builder, Customize(identifier));
 }
